RUB-250: split Go P2P wire helpers
